auth is now optional

+++ src/entrypoint.sh
@@ -13,15 +13,20 @@ fi
 
 > /etc/nginx/.htpasswd
 
-echo "Generation of .htpassword"
-for i in ${LST_USERS_PASSWORDS}
-do
-    user=$(echo $i | cut -d: -f1)
-    password=$(echo $i | cut -d: -f2)
-    #echo "$user:$password" >> /etc/nginx/.htpasswd
-    htpasswd -b -c /etc/nginx/.htpasswd.$$  $user $password
-    cat /etc/nginx/.htpasswd.$$ >> /etc/nginx/.htpasswd
-    rm /etc/nginx/.htpasswd.$$
-done
+if [ ! -z "$LST_USERS_PASSWORDS" ]
+then
+    echo "Generation of .htpassword"
+    for i in ${LST_USERS_PASSWORDS}
+    do
+        user=$(echo $i | cut -d: -f1)
+        password=$(echo $i | cut -d: -f2)
+        #echo "$user:$password" >> /etc/nginx/.htpasswd
+        htpasswd -b -c /etc/nginx/.htpasswd.$$  $user $password
+        cat /etc/nginx/.htpasswd.$$ >> /etc/nginx/.htpasswd
+        rm /etc/nginx/.htpasswd.$$
+    done
+else
+    sed -i '/auth_basic/d' /etc/nginx/nginx.conf
+fi
 
 exec $@
